chore: better formatting of the enterprise api views
