+++ CHANGELOG.rst
@@ -15,6 +15,10 @@ Change Log
 
 Unreleased
 ----------
+[3.67.5]
+--------
+chore: better formatting of the enterprise api views
+
 [3.67.4]
 --------
 feat: add button to update customer modified time
