+++ enterprise/api/v1/views/__init__.py
@@ -1,0 +1,3 @@
+"""
+API views for the enterprise app.
+"""
