+++ enterprise/api/v1/views/enterprise_customer_branding_configuration.py
@@ -1,7 +1,6 @@
 """
 Views for the ``enterprise-customer-branding`` API endpoint.
 """
-from logging import getLogger
 
 from edx_rbac.decorators import permission_required
 from rest_framework import permissions, status
@@ -9,11 +8,12 @@ from rest_framework.decorators import action
 from rest_framework.parsers import FormParser, MultiPartParser
 from rest_framework.response import Response
 
-from enterprise.api.v1.views.base_views import EnterpriseReadWriteModelViewSet
 from enterprise import models
 from enterprise.api.v1 import serializers
+from enterprise.api.v1.views.base_views import EnterpriseReadWriteModelViewSet
+from enterprise.logging import getEnterpriseLogger
 
-LOGGER = getLogger(__name__)
+LOGGER = getEnterpriseLogger(__name__)
 
 
 class EnterpriseCustomerBrandingConfigurationViewSet(EnterpriseReadWriteModelViewSet):
